+++ amplicode.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AmplicodePersistenceUnitSettings">
+    <persistence-units>
+      <persistence-unit moduleName="product-service-app" name="Default">
+        <packages>
+          <package value="com.masya.product_service" />
+        </packages>
+      </persistence-unit>
+    </persistence-units>
+  </component>
+</project>
