--- product-service-app/src/main/java/com/iprody/product_service/domain/exception/ResourceNotFoundException.java
@@ -1,8 +1,0 @@
-package com.iprody.product_service.domain.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
