--- product-service-app/src/main/java/com/iprody/product_service/domain/model/Product.java
@@ -1,71 +1,0 @@
-package com.iprody.product_service.domain.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.math.BigDecimal;
-import java.time.Instant;
-
-@Entity
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "product")
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "product_seq")
-    @SequenceGenerator(name = "product_seq", sequenceName = "product_sequence", allocationSize = 1)
-    private Long id;
-
-    @Column(name = "summary", nullable = false, length = 100)
-    private String summary;
-
-    @Column(name = "description", nullable = false)
-    private String description;
-
-    @Column(name = "price", nullable = false)
-    private BigDecimal price;
-
-    @Column(name = "duration", nullable = false)
-    private Short duration;
-
-    @ManyToOne
-    @JoinColumn(name = "discount_id")
-    private Discount discount;
-
-    @Column(name = "active", nullable = false)
-    private boolean active;
-
-    @CreationTimestamp
-    @Column(name = "created_at", nullable = false)
-    private Instant createdAt;
-
-    @CreationTimestamp
-    @Column(name = "updated_at", nullable = false)
-    private Instant updatedAt;
-
-    public Product(String summary, String description, BigDecimal price, Short duration, Discount discount, boolean active, Instant createdAt, Instant updatedAt) {
-        this.summary = summary;
-        this.description = description;
-        this.price = price;
-        this.duration = duration;
-        this.discount = discount;
-        this.active = active;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-}
