--- product-service-app/src/main/java/com/iprody/product_service/domain/service/DiscountService.java
@@ -1,21 +1,0 @@
-package com.iprody.product_service.domain.service;
-
-
-import com.iprody.product_service.web.dto.DiscountDto;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface DiscountService {
-
-    List<DiscountDto> getAllDiscounts();
-
-    DiscountDto createDiscount(DiscountDto discountDto);
-
-    void updateDiscount(Long id, DiscountDto discountDto);
-
-    Optional<DiscountDto> getDiscountById(Long id);
-
-    void deleteDiscount(Long id);
-
-}
