--- product-service-app/src/main/java/com/iprody/product_service/web/dto/DiscountDto.java
@@ -1,31 +1,0 @@
-package com.iprody.product_service.web.dto;
-
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.Instant;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class DiscountDto {
-
-    private Long id;
-
-    @Min(0)
-    @Max(100)
-    private Short amount;
-
-    @NotNull
-    private Instant startTime;
-
-    @NotNull
-    private Instant endTime;
-
-}
