--- product-service-app/src/main/java/com/iprody/product_service/web/dto/ProductDto.java
@@ -1,42 +1,0 @@
-package com.iprody.product_service.web.dto;
-
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProductDto {
-
-    private Long id;
-
-    @NotBlank
-    @Size(max = 100)
-    private String summary;
-
-    @NotBlank
-    private String description;
-
-    @NotNull
-    @DecimalMin("0.0")
-    private BigDecimal price;
-
-    @NotNull
-    private Short duration;
-
-    @NotNull
-    private Long discountId;
-
-    @NotNull
-    private boolean active;
-
-}
