--- product-service-app/src/main/java/com/iprody/product_service/web/dto/mapper/Mappable.java
@@ -1,9 +1,0 @@
-package com.iprody.product_service.web.dto.mapper;
-
-public interface Mappable<E, D> {
-
-    E toEntity(D dto);
-
-    D toDto(E entity);
-
-}
