+++ product-service-app/src/main/java/com/masya/product_service/ProductServiceApplication.java
@@ -1,0 +1,13 @@
+package com.masya.product_service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProductServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProductServiceApplication.class, args);
+    }
+
+}
