+++ product-service-app/src/main/java/com/masya/product_service/domain/exception/ResourceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.masya.product_service.domain.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
