+++ product-service-app/src/main/java/com/masya/product_service/domain/model/Discount.java
@@ -1,0 +1,61 @@
+package com.masya.product_service.domain.model;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.Instant;
+import java.util.List;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "discount")
+public class Discount {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "discount_seq")
+    @SequenceGenerator(name = "discount_seq", sequenceName = "discount_sequence", allocationSize = 1)
+    private Long id;
+
+    @Column(name = "amount")
+    private Short amount;
+
+    @Column(name = "start_time")
+    private Instant startTime;
+
+    @Column(name = "end_time")
+    private Instant endTime;
+
+    @CreationTimestamp
+    @Column(name = "created_at")
+    private Instant createdAt;
+
+    @CreationTimestamp
+    @Column(name = "updated_at")
+    private Instant updatedAt;
+
+    @OneToMany(mappedBy = "discount")
+    private List<Product> products;
+
+    public Discount(Short amount, Instant startTime, Instant endTime, Instant createdAt, Instant updatedAt) {
+        this.amount = amount;
+        this.startTime = startTime;
+        this.endTime = endTime;
+        this.createdAt = createdAt;
+        this.updatedAt = updatedAt;
+    }
+}
+
