+++ product-service-app/src/main/java/com/masya/product_service/domain/repository/DiscountRepository.java
@@ -1,0 +1,7 @@
+package com.masya.product_service.domain.repository;
+
+import com.masya.product_service.domain.model.Discount;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DiscountRepository extends JpaRepository<Discount, Long> {
+}
