+++ product-service-app/src/main/java/com/masya/product_service/web/dto/mapper/Mappable.java
@@ -1,0 +1,9 @@
+package com.masya.product_service.web.dto.mapper;
+
+public interface Mappable<E, D> {
+
+    E toEntity(D dto);
+
+    D toDto(E entity);
+
+}
