+++ product-service-app/src/main/java/com/masya/product_service/web/dto/mapper/ProductMapper.java
@@ -1,0 +1,21 @@
+package com.masya.product_service.web.dto.mapper;
+
+import com.masya.product_service.domain.model.Product;
+import com.masya.product_service.web.dto.ProductDto;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
+@Mapper(componentModel = "spring", uses = {DiscountMapper.class})
+public interface ProductMapper extends Mappable<Product, ProductDto> {
+
+    @Override
+    @Mapping(target = "updatedAt", ignore = true)
+    @Mapping(target = "createdAt", ignore = true)
+    @Mapping(target = "discount.id", source = "discountId")
+    Product toEntity(ProductDto dto);
+
+    @Override
+    @Mapping(target = "discountId", source = "discount.id")
+    ProductDto toDto(Product entity);
+
+}
