--- product-service-app/src/test/java/com/iprody/product_service/utils/DiscountBuilder.java
@@ -1,18 +1,0 @@
-package com.iprody.product_service.utils;
-
-import com.iprody.product_service.domain.model.Discount;
-
-import java.time.Instant;
-
-public class DiscountBuilder {
-
-    public static Discount buildDiscount() {
-        return new Discount(
-                (short) 1,
-                Instant.now(),
-                Instant.now(),
-                Instant.now(),
-                Instant.now());
-    }
-
-}
