+++ product-service-app/src/test/java/com/masya/product_service/utils/DiscountBuilder.java
@@ -1,0 +1,18 @@
+package com.masya.product_service.utils;
+
+import com.masya.product_service.domain.model.Discount;
+
+import java.time.Instant;
+
+public class DiscountBuilder {
+
+    public static Discount buildDiscount() {
+        return new Discount(
+                (short) 1,
+                Instant.now(),
+                Instant.now(),
+                Instant.now(),
+                Instant.now());
+    }
+
+}
